+++ packages/client/test/integration/Sequencing.test.ts
@@ -8,17 +8,18 @@ import { collect } from '../../src/utils/iterators'
 import { FakeEnvironment } from '../test-utils/fake/FakeEnvironment'
 import { getWaitForStorage } from '../test-utils/publish'
 import { createTestStream, uid } from '../test-utils/utils'
+import { Message } from '../../src/Message'
 
 const Msg = (opts?: any) => ({
     value: uid('msg'),
     ...opts,
 })
 
-function toSeq(requests: any[], ts = Date.now()) {
-    return requests.map((streamMessage) => {
-        const { prevMsgRef } = streamMessage
+function toSeq(requests: Message[], ts = Date.now()) {
+    return requests.map((msg) => {
+        const { prevMsgRef } = msg.streamMessage
         return [
-            [streamMessage.getTimestamp() - ts, streamMessage.getSequenceNumber()],
+            [msg.timestamp - ts, msg.sequenceNumber],
             prevMsgRef ? [prevMsgRef.timestamp - ts, prevMsgRef.sequenceNumber] : null
         ]
     })
@@ -165,7 +166,7 @@ describe('Sequencing', () => {
                 }
             }
         )
-        const msgsResent = (await collect(sub)).map((m) => m.getParsedContent())
+        const msgsResent = (await collect(sub)).map((m) => m.content)
 
         expect(msgsReceieved).toEqual(msgsResent)
         // backdated messages disappear
@@ -233,7 +234,7 @@ describe('Sequencing', () => {
                 }
             }
         )
-        const msgsResent = (await collect(sub)).map((m) => m.getParsedContent())
+        const msgsResent = (await collect(sub)).map((m) => m.content)
 
         expect(msgsReceieved).toEqual(msgsResent)
         // backdated messages disappear