+++ packages/client/test/integration/resend-and-subscribe.test.ts
@@ -83,14 +83,14 @@ describe('resend and subscribe', () => {
         await publisher.publish(stream.id, { mockId: 2 }, { timestamp: 2000 })
 
         const receivedMessage2 = await nextValue(messageIterator)
-        expect(receivedMessage1!.getParsedContent()).toEqual({
+        expect(receivedMessage1!.content).toEqual({
             mockId: 1
         })
-        expect(receivedMessage1!.groupKeyId).toBe(groupKey.id)
-        expect(receivedMessage2!.getParsedContent()).toEqual({
+        expect(receivedMessage1!.streamMessage.groupKeyId).toBe(groupKey.id)
+        expect(receivedMessage2!.content).toEqual({
             mockId: 2
         })
-        expect(receivedMessage2!.groupKeyId).toBe(groupKey.id)
+        expect(receivedMessage2!.streamMessage.groupKeyId).toBe(groupKey.id)
         const groupKeyRequests = environment.getNetwork().getSentMessages({
             messageType: StreamMessage.MESSAGE_TYPES.GROUP_KEY_REQUEST
         })